requirements.txt
mcp==1.6.0
httpx==0.28.1
python-dotenv==1.1.0
uvicorn==0.34.1
